// src/types/colorTypes.ts
enum ColorType {
  BACKGROUND = "background",
  BORDER = "border",
}

enum ColorGroupNameRU {
  BACKGROUND = "Основной цвет",
  BORDER = "Цвет границы",
}

enum ColorLabelStyle {
  BACKGROUND = "color__label--background",
  BORDER = "color__label--border",
}

enum ColorButtonStyleSelected {
  BACKGROUND = "color__selected--background",
  BORDER = "color__selected--border",
}

export { ColorType, ColorGroupNameRU, ColorButtonStyleSelected, ColorLabelStyle };